// src/redux/filter/types.ts
export enum SortPropertyEnum {
    RATING = 'rating',
    TITLE = 'title',
    PRICE = 'price'
  }
  
  export type Sort = {
    name: string;
    sortProperty: SortPropertyEnum;
  }
  
  export interface FilterSliceState {
    searchValue: string;
    categoryId: number;
    currentPage: number;
    sort: Sort;
  }